% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rufosightings}
\alias{Rufosightings}
\title{UFO Sightings}
\format{A data frame with 55633 rows and 9 variables:
\describe{
  item{sighted_at_year}{sighted_at_year}
  item{sighted_at_month}{sighted_at_month}
  item{sighted_at_day}{sighted_at_day}
  item{reported_at}{reported_at}
  item{location_city}{location_city}
  item{location_state}{location_state}
  item{shape}{shape}
  item{duration}{duration}
  item{description}{description}
}}
\source{
\url{http://shop.oreilly.com/product/0636920018483.do}
}
\usage{
Rufosightings
}
\description{
A dataset containing reports of UFO sightings
}
\keyword{datasets}

