% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ufo_byexactdate.R
\name{ufo_byexactdate}
\alias{ufo_byexactdate}
\title{ufo_byexactdate function}
\usage{
ufo_byexactdate(exactdate = 19951011)
}
\arguments{
\item{exactdata}{Defaults to 19951011
ufo_byexactdate()}
}
\description{
This function allows you to find UFO sightings on an exact date.
}

